refactor: extract notes data layer into useNotes hook and split out NoteEditor
Split web/src/pages/Notes.tsx into three units:
- web/src/hooks/useNotes.ts owns list/tag fetching (with AbortController
  cleanup) plus save/remove mutations and refresh, exposing the shared
  Note/ViewMode types.
- web/src/components/notes/NoteEditor.tsx owns draft state, edit/preview
  toggling, the dirty check, and the ReactMarkdown component map.
- Notes.tsx is now a thin layout shell (list, search, tag filters,
  selection, delete confirm dialog) that wires the hook to the editor.

Pure structural refactor: no behavioral or API changes. Added a wiring
test covering load, tag filtering, dirty-gated save, create POST, and the
delete confirmation/DELETE flow.

Bead: Hytte-62fw
